+++ src/com/iTeam/controller/GoodsTypeController.java
@@ -129,6 +129,8 @@ public class GoodsTypeController {
 	 */
 	@RequestMapping(value = "/goodsType",method = RequestMethod.DELETE,produces = PRODUCES)
 	public MyResponse deleteType(@RequestBody List<Integer> ids) throws IOException {
+		if(ids.isEmpty())
+			return new MyResponse().failure("要删除的数目为零");
 		for(int i = 0;i < ids.size();i++) {
 			service.delete(ids.get(i));
 		}
