+++ src/com/iTeam/controller/StorageController.java
@@ -91,6 +91,8 @@ public class StorageController {
 	//在浏览中通过URL调用deleteStorage这个方法
 	@RequestMapping(value="/storage",method=RequestMethod.DELETE,produces=PRODUCES)
 	public MyResponse deleteStorage(@RequestBody List<Integer> ids)throws Exception{
+		if(ids.isEmpty())
+			return new MyResponse().failure("要删除的数目为零");
 		for(int i=0;i<ids.size();i++){
 			service.delete(ids.get(i));
 			System.out.println(ids.get(i));
