+++ src/com/iTeam/controller/UserMessageController.java
@@ -158,6 +158,8 @@ public class UserMessageController {
 	 */
 	@RequestMapping(value="/user",method=RequestMethod.DELETE,produces = PRODUCES)
 	public MyResponse delete(@RequestBody List<Integer> ids) throws IOException {
+		if(ids.isEmpty())
+			return new MyResponse().failure("要删除的数目为零");
 		//支持删除多条记录
 		for(int i = 0;i < ids.size();i++) {
 			userMessageService.delete(ids.get(i));
